addition of error pages & blog comments

+++ blog/views.py
@@ -1,15 +1,21 @@
-from django.shortcuts import render, get_object_or_404, redirect
+from django.shortcuts import render, get_object_or_404, redirect, render_to_response
 from django.contrib.auth.decorators import login_required
+from django.views.decorators.csrf import requires_csrf_token
+from django.http import HttpResponseForbidden, HttpResponseRedirect
 from blog.models import Post, Comment
 from django.utils import timezone
+from django.core.paginator import Paginator
 from blog.forms import PostForm, CommentForm
-
+from django.contrib import messages
 from django.views.generic import (TemplateView,ListView,
                                   DetailView,CreateView,
                                   UpdateView,DeleteView)
-
+from django.template import RequestContext
 from django.urls import reverse_lazy
 from django.contrib.auth.mixins import LoginRequiredMixin
+from django.views.generic import RedirectView
+from django.contrib.auth.decorators import login_required
+from .models import Post
 
 class PostListView(ListView):
     model = Post
@@ -32,14 +38,6 @@ class CreatePostView(LoginRequiredMixin,CreateView):
     model = Post
 
 
-class PostUpdateView(LoginRequiredMixin,UpdateView):
-    login_url = '/login/'
-    redirect_field_name = 'blog/post_detail.html'
-
-    form_class = PostForm
-
-    model = Post
-
 
 class DraftListView(LoginRequiredMixin,ListView):
     login_url = '/login/'
@@ -50,10 +48,38 @@ class DraftListView(LoginRequiredMixin,ListView):
     def get_queryset(self):
         return Post.objects.filter(published_date__isnull=True).order_by('created_date')
 
-
-class PostDeleteView(LoginRequiredMixin,DeleteView):
-    model = Post
-    success_url = reverse_lazy('post_list')
+@requires_csrf_token  
+def PostDeleteView(request, pk):
+    post = get_object_or_404(Post, pk=pk)
+    
+    if not(request.user == post.author or request.user.is_superuser):
+        return render(request, '403.html')
+    
+    if request.method == "POST":
+        form = PostForm(request.POST, request.FILES, instance=post)
+        if form.is_valid():
+            post = form.save()
+            return redirect('post_detail', post.pk)        
+    else:
+        form = PostForm(instance=post)
+    return render(request, 'blog/post_form.html', {'form': form})
+    
+  
+@requires_csrf_token 
+def PostUpdateView(request, pk):
+    post = get_object_or_404(Post, pk=pk)
+    
+    if not(request.user == post.author or request.user.is_superuser):
+        return render(request, "403.html")
+    
+    if request.method == "POST":
+        form = PostForm(request.POST, request.FILES, instance=post)
+        if form.is_valid():
+            post = form.save()
+            return redirect('post_detail', post.pk)        
+    else:
+        form = PostForm(instance=post)
+    return render(request, 'blog/post_form.html', {'form': form})
 
 #######################################
 ## Functions that require a pk match ##
@@ -74,6 +100,7 @@ def add_comment_to_post(request, pk):
             comment = form.save(commit=False)
             comment.post = post
             comment.save()
+            messages.success(request, "Thanks for your feedback. Your comment will appear once approved")
             return redirect('post_detail', pk=post.pk)
     else:
         form = CommentForm()
@@ -83,13 +110,21 @@ def add_comment_to_post(request, pk):
 @login_required
 def comment_approve(request, pk):
     comment = get_object_or_404(Comment, pk=pk)
-    comment.approve()
+    if not request.user.is_superuser :
+        return HttpResponseForbidden('You do not have permission to approve!')
+    else:
+        comment.approve()
     return redirect('post_detail', pk=comment.post.pk)
+
 
 
 @login_required
 def comment_remove(request, pk):
     comment = get_object_or_404(Comment, pk=pk)
     post_pk = comment.post.pk
-    comment.delete()
+    if not request.user.is_superuser :
+        return HttpResponseForbidden('You do not have permission to delete!')
+    else:
+        comment.delete()
     return redirect('post_detail', pk=post_pk)
+
